requirements.txt
langchain-google-genai
langchain
python-dotenv
faiss-cpu
langchain-community
pypdf
langchain-text-splitters
